Making MAKE_FEASIBLE function.
**Problem in GIT, master not showing files. (Solved, First change to master, then merged 1st_branch to HEAD)

+++ src/CTSP/FITNESS.java
@@ -22,7 +22,7 @@ class FITNESS {
         }
 
         for (int i = 0; i < popSize; i++) {   // For putting max-fitness, for roulette whel selection
-            Array[i][noc + 1] = (int) max - Array[i][noc];
+            Array[i][noc + 1] = max - Array[i][noc];
 //            System.out.println(popArray[i][noc + 1]); // Debug Only
         }
